requirements.txt
argparse==1.4.0
pyyaml=5.1.2
slackclient = 1.3.2
pytest=5.1.0
nbval==0.9.2
patsy==0.5.1
